// prisma/src/generated/prisma/models/AveragePrices.ts
export type AveragePricesCreateInput = {
  id?: string
  brand: string
  model: string
  avg_price: runtime.Decimal | runtime.DecimalJsLike | number | string
  cars_count: number
  calculated_at?: Date | string
  Region?: Prisma.RegionCreateNestedOneWithoutAveragePricesInput
}